requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
pymysql==1.1.0